Tests/LibWeb/Text/expected/background-position-xy.txt

right 0px
bottom 0px
